requirements.txt
numpy==1.23.3
keras
tensorflow==2.10.0
Pillow
pandas
matplotlib
sklearn
tqdm
ipynb